Hard-code collections hubs for collect page to avoid metaphysics problem
