+++ src/Components/__stories__/ImageLink.story.tsx
@@ -1,5 +1,6 @@
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
+import { imageSamples } from "../_fixtures_/collectionsHubs"
 import { ImageLink } from "../ImageLink"
 
 storiesOf("Components/ImageLink", module)
@@ -38,18 +39,3 @@ storiesOf("Components/ImageLink", module)
       Impressionist and Modern Art
     </ImageLink>
   ))
-
-export const imageSamples = {
-  contemporaryArt:
-    "https://d32dm0rphc51dk.cloudfront.net/f_WVnADS9HIc5dQ-sIcejA/thumb.jpg",
-  postWarArt:
-    "https://d32dm0rphc51dk.cloudfront.net/dtSncXEq-KNTWbWNG_xMTA/thumb.jpg",
-  impressionistAndModernArt:
-    "https://d32dm0rphc51dk.cloudfront.net/Y2fVKtk64zRDfoGWgYSkJA/thumb.jpg",
-  preTwentiethCentury:
-    "https://d32dm0rphc51dk.cloudfront.net/adz_7LkzkU5A_ucVjQLMtQ/thumb.jpg",
-  photography:
-    "https://d32dm0rphc51dk.cloudfront.net/Cy1tDMUKkF_H-QN4BIDlDA/thumb.jpg",
-  streetArt:
-    "https://d32dm0rphc51dk.cloudfront.net/Tk7srLDTS-0Y60mbN7gWew/thumb.jpg",
-}
